chore: extract common parts to base controller
