+++ packages/shared/widget-plugin-filtering/src/controllers/picker/StaticComboboxController.ts
@@ -1,58 +1,28 @@
 import { useCombobox, UseComboboxProps } from "downshift";
-import { ActionValue, DynamicValue, EditableValue } from "mendix";
-import { autorun, makeAutoObservable, observable, reaction } from "mobx";
+import { action, autorun, makeObservable, observable, reaction } from "mobx";
 import { disposeFx } from "../../mobx-utils";
-import { OptionsSerializer } from "../../stores/picker/OptionsSerializer";
-import { StaticSelectFilterStore } from "../../stores/picker/StaticSelectFilterStore";
-import { ResetHandler, SetValueHandler } from "../../typings/IJSActionsControlled";
 import { OptionWithState } from "../../typings/OptionWithState";
-import { PickerChangeHelper } from "../generic/PickerChangeHelper";
-import { PickerJSActionsHelper } from "../generic/PickerJSActionsHelper";
+import { StaticBaseController, StaticBaseControllerProps } from "./StaticBaseController";
 
-interface Props {
-    defaultValue?: string;
-    filterOptions: Array<CustomOption<DynamicValue<string>>>;
-    filterStore: StaticSelectFilterStore;
-    onChange?: ActionValue;
-    valueAttribute?: EditableValue<string>;
+interface Props extends StaticBaseControllerProps {
     inputPlaceholder?: string;
 }
 
-interface CustomOption<T> {
-    caption: T;
-    value: T;
-}
-
-export class StaticComboboxController {
-    private actionHelper: PickerJSActionsHelper;
-    private changeHelper: PickerChangeHelper;
-    private defaultValue?: Iterable<string>;
-    private filterStore: StaticSelectFilterStore;
-    private serializer: OptionsSerializer;
+export class StaticComboboxController extends StaticBaseController {
     private touched = false;
-
-    filterOptions: Array<CustomOption<DynamicValue<string>>>;
     inputValue: string;
     inputPlaceholder: string;
 
     constructor(props: Props) {
+        super({ ...props, multiselect: false });
         this.inputPlaceholder = props.inputPlaceholder ?? "Search";
-        this.filterOptions = props.filterOptions;
-        this.filterStore = props.filterStore;
-        this.serializer = new OptionsSerializer({ store: this.filterStore });
-        this.defaultValue = this.serializer.fromStorableValue(props.defaultValue);
-        this.actionHelper = new PickerJSActionsHelper({
-            filterStore: props.filterStore,
-            parse: value => this.serializer.fromStorableValue(value) ?? [],
-            multiselect: false
-        });
-
-        this.changeHelper = new PickerChangeHelper(props, () => this.serializer.value);
         this.inputValue = this.selectedOption?.caption ?? "";
-        makeAutoObservable(this, {
-            useComboboxProps: false,
-            setup: false,
-            filterOptions: observable.struct
+
+        makeObservable<this, "touched">(this, {
+            inputValue: observable,
+            setInputValue: action,
+            touched: observable,
+            setTouched: action
         });
     }
 
@@ -68,35 +38,17 @@ export class StaticComboboxController {
                 }
             })
         );
+
+        // Set input when selected option changes
         disposers.push(
             reaction(
                 (): string => this.selectedOption?.caption ?? "",
                 caption => this.setInputValue(caption)
             )
         );
-        disposers.push(
-            autorun(() => {
-                if (this.filterOptions.length > 0) {
-                    const options = this.filterOptions.map(this.toStoreOption);
-                    this.filterStore.setCustomOptions(options);
-                }
-            })
-        );
-        disposers.push(this.changeHelper.setup());
 
-        // Set default after all reactions are set up
-        if (this.defaultValue) {
-            this.filterStore.setDefaultSelected(this.defaultValue);
-        }
+        disposers.push(super.setup());
         return dispose;
-    }
-
-    updateProps(props: Props): void {
-        this.changeHelper.updateProps(props);
-    }
-
-    get options(): OptionWithState[] {
-        return this.filterStore.options;
     }
 
     get selectedIndex(): number {
@@ -107,15 +59,6 @@ export class StaticComboboxController {
     get selectedOption(): OptionWithState | null {
         return this.filterStore.allOptions.find(option => option.selected) || null;
     }
-
-    get isEmpty(): boolean {
-        return this.filterStore.selected.size === 0;
-    }
-
-    toStoreOption = (opt: CustomOption<DynamicValue<string>>): CustomOption<string> => ({
-        caption: `${opt.caption?.value}`,
-        value: `${opt.value?.value}`
-    });
 
     setTouched(value: boolean): void {
         this.touched = value;
@@ -139,14 +82,6 @@ export class StaticComboboxController {
         this.setTouched(false);
         this.setInputValue("");
         this.filterStore.clear();
-    };
-
-    handleSetValue = (...args: Parameters<SetValueHandler>): void => {
-        this.actionHelper.handleSetValue(...args);
-    };
-
-    handleResetValue = (...args: Parameters<ResetHandler>): void => {
-        this.actionHelper.handleResetValue(...args);
     };
 
     useComboboxProps = (): UseComboboxProps<OptionWithState> => {
